+++ src/app/static/src/app/store/root/getters.ts
@@ -1,7 +1,7 @@
 import {RootState} from "../../root";
 import {Getter, GetterTree} from "vuex";
 import {Warning} from "../../generated";
-import {Dict} from "../../types";
+import {Dict, StepWarnings} from "../../types";
 
 interface RootGetters {
     isGuest: Getter<RootState, RootState>
@@ -21,11 +21,13 @@ export const getters: RootGetters & GetterTree<RootState, RootState> = {
         return state.currentUser == "guest";
     },
 
-    warnings: (state: RootState) => (stepName: string) => {
+    warnings: (state: RootState) => (stepName: string): StepWarnings => {
         const filterWarnings = (warnings: Warning[], stepLocation: string) =>
-            warnings.filter(warning => warning.locations.some(location => location === stepLocation))
+            stepLocation ?
+                (warnings || []).filter(warning => warning.locations.some(location => location === stepLocation)) :
+                [];
 
-        const location = warningStepLocationMapping[stepName]
+        const location = warningStepLocationMapping[stepName];
 
         return {
             modelOptions: filterWarnings(state.modelOptions.warnings, location),