+++ src/app/static/src/tests/setup.ts
@@ -24,7 +24,8 @@ i18next.init({
     lng: Language.en,
     resources: {
         en: {translation: locales.en},
-        fr: {translation: locales.fr}
+        fr: {translation: locales.fr},
+        pt: {translation: locales.pt}
     },
     fallbackLng: Language.en
 });